community: make room for directed Louvain implementation
